refactoring of bundle
